rewrite bindings
it seems easier to manually copy them from the DYNALIB_FN declarations in the
spark/firmware repo rather than running bindgen.
